+++ lib/last_crusader/micropub/hugo.ex
@@ -141,6 +141,7 @@ defmodule LastCrusader.Micropub.Hugo do
     text
     |> RemoveEmoji.sanitize()
     |> String.replace("—", "-")
+    |> String.replace("–", "-")
     |> Slugger.slugify_downcase()
     |> Slugger.truncate_slug(31)
   end